+++ server/main.ts
@@ -1,17 +1,26 @@
+import type { NextFunction, Request, Response } from 'express'
+
 const path = require('path')
 const fs = require('fs').promises
 
+interface RegisterServerOptions {
+  registerHook: any
+  registerSetting: any
+  settingsManager: any
+  storageManager: any
+  videoCategoryManager: any
+  videoLicenceManager: any
+  videoLanguageManager: any
+  getRouter: any
+  peertubeHelpers: any
+}
+
 async function register ({
-  _registerHook,
   registerSetting,
   settingsManager,
-  _storageManager,
-  _videoCategoryManager,
-  _videoLicenceManager,
-  _videoLanguageManager,
   getRouter,
   peertubeHelpers
-}) {
+}: RegisterServerOptions): Promise<any> {
   registerSetting({
     name: 'chat-auto-display',
     label: 'Automatically open the chat',
@@ -142,8 +151,8 @@ async function register ({
   const converseJSIndex = await fs.readFile(path.resolve(__dirname, './conversejs/index.html'))
 
   const router = getRouter()
-  router.get('/ping', (req, res) => res.json({ message: 'pong' }))
-  router.get('/webchat', async (req, res, next) => {
+  router.get('/ping', (req: Request, res: Response) => res.json({ message: 'pong' }))
+  router.get('/webchat', async (req: Request, res: Response, next: NextFunction) => {
     try {
       const settings = await settingsManager.getSettings([
         'chat-use-builtin', 'chat-room', 'chat-server',
@@ -168,7 +177,7 @@ async function register ({
       // be /webchat/:videoId
       // const id = req.param('videoId')
       // const video = await peertubeHelpers.videos.loadByIdOrUUID(id)
-      let url = req.query.url
+      let url: string = req.query.url as string || ''
       if (!url) {
         throw new Error('Missing url parameter)')
       }
@@ -183,7 +192,7 @@ async function register ({
         throw new Error('Video not found')
       }
 
-      let page = '' + converseJSIndex
+      let page = '' + (converseJSIndex as string)
       // FIXME: Peertube should provide the static folder path. For now:
       const staticRelative = '../static'
       page = page.replace(/{{BASE_STATIC_URL}}/g, staticRelative)
@@ -202,7 +211,7 @@ async function register ({
   })
 }
 
-async function unregister () {
+async function unregister (): Promise<any> {
 }
 
 module.exports = {